+++ semcore/d3-chart/src/StackBar.jsx
@@ -4,7 +4,7 @@ import { Component } from '@semcore/core';
 import getOriginChildren from '@semcore/utils/lib/getOriginChildren';
 import createElement from './createElement';
 import Bar, { MIN_HEIGHT } from './Bar';
-import HorizontalBar from './HorizontalBar';
+import HorizontalBar, { MIN_WIDTH } from './HorizontalBar';
 
 const DEFAULT_INSTANCE = Symbol('DEFAULT_INSTANCE');
 const XY0 = Symbol('XY0');
@@ -69,7 +69,7 @@ class StackBarRoot extends Component {
         yScale(d[y]) - Math.min(yScale(yScale.domain()[0]), yScale(d[XY0] ?? 0)),
       );
       this.offsetBars[seriesIndex][i] =
-        Number(d[y] - (d[XY0] ?? 0)) === 0 ? 0 : absHeight >= hMin ? 0 : hMin;
+        Number(d[y] - (d[XY0] ?? 0)) === 0 ? 0 : absHeight >= hMin ? 0 : d[y] > 0 ? hMin : -hMin;
       return [0, offset];
     };
 
@@ -84,8 +84,9 @@ class StackBarRoot extends Component {
     };
   }
 
-  getHorizontalBarProps({ x }) {
-    const { y, r } = this.asProps;
+  getHorizontalBarProps({ x, wMin = MIN_WIDTH }) {
+    const { y, r, scale } = this.asProps;
+    const [xScale] = scale;
 
     const seriesIndex = this.series.findIndex((s) => s.key === x);
     const series = this.series[seriesIndex];
@@ -94,17 +95,32 @@ class StackBarRoot extends Component {
       this.series.slice(seriesIndex + 1).some((bar) => bar[i][0] !== bar[i][1]) ? 0 : r,
     );
 
+    this.offsetBars[seriesIndex] = this.offsetBars[seriesIndex] ?? [];
+    const data = series.map((s) => ({
+      ...s.data,
+      [x]: s[1],
+      [XY0]: s[0],
+    }));
+
+    const calcOffset = (i) => {
+      const offset = this.offsetBars.reduce((offset, offsetBar) => offset - (offsetBar[i] ?? 0), 0);
+      const d = data[i];
+      const absWidth = Math.abs(
+        xScale(d[x]) - Math.max(xScale(xScale.domain()[0]), xScale(d[XY0] ?? 0)),
+      );
+      this.offsetBars[seriesIndex][i] =
+        Number(d[x] - (d[XY0] ?? 0)) === 0 ? 0 : absWidth >= wMin ? 0 : d[x] > 0 ? -wMin : wMin;
+      return [offset, 0];
+    };
+
     return {
-      data: series.map((s) => ({
-        ...s.data,
-        [x]: s[1],
-        [XY0]: s[0],
-      })),
-      wMin: 0,
+      data,
+      wMin,
       x0: XY0,
       y,
       r: rBar,
       groupKey: y,
+      offset: calcOffset,
     };
   }